Adjust various chart-related units for better precision/intent

+++ YARG.Core/Chart/Events/WaitCountdown.cs
@@ -5,8 +5,8 @@ namespace YARG.Core.Chart
 {
     public class WaitCountdown : ChartEvent
     {
-        public const float MIN_SECONDS = 9;
-        public const float END_COUNTDOWN_SECOND = 1f;
+        public const double MIN_SECONDS = 9;
+        public const double END_COUNTDOWN_SECOND = 1;
 
         //The time where the countdown should start fading out and overstrums will break combo again
         public double DeactivateTime => TimeEnd - END_COUNTDOWN_SECOND;
